feat: Add logging settings, merge ScannerTesterActivity into the MainActivity
