+++ demoscannerapp/src/main/res/layout/activity_settings.xml
@@ -274,6 +274,37 @@
                     android:layout_marginEnd="16dp"/>
             </LinearLayout>
 
+            <LinearLayout
+                android:id="@+id/linearLayoutEnableLogging"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal"
+                android:gravity="center_vertical"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@+id/linearLayoutIntentDevice">
+
+                <TextView
+                    android:id="@+id/textViewEnableLoging"
+                    android:layout_width="0dp"
+                    android:layout_weight="9"
+                    android:layout_height="wrap_content"
+                    android:text="@string/enable_log"
+                    android:textAppearance="?attr/textAppearanceBody2"
+                    android:textColor="@color/defaultTextColor"
+                    android:layout_marginStart="16dp"/>
+
+                <View
+                    android:layout_width="0dp"
+                    android:layout_height="0dp"
+                    android:layout_weight="1" />
+
+                <com.google.android.material.materialswitch.MaterialSwitch
+                    android:id="@+id/switchEnableLogging"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginEnd="16dp"/>
+            </LinearLayout>
+
             <com.google.android.material.divider.MaterialDivider
                 android:id="@+id/dividerSettingsAllowedProvider"
                 android:layout_width="0dp"
@@ -282,7 +313,7 @@
                 app:layout_constraintWidth_percent="0.90"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/linearLayoutIntentDevice"/>
+                app:layout_constraintTop_toBottomOf="@+id/linearLayoutEnableLogging"/>
 
             <!-- ALLOWED PROVIDERS -->
